fix(sql): cast count(*) to int in the RFM backfill
count(*) คืน bigint → Postgres หา fn_rfm_segment(int, timestamptz) ไม่เจอ
และ DO block ล้มทั้งก้อน (42883) ทำให้ส่วนที่เหลือของไฟล์ไม่ได้รัน

+++ sql/harden_loyalty_integrity.sql
@@ -81,11 +81,12 @@ begin
     alter table public.customers disable trigger tr_customers_guard_points;
   end if;
 
+  -- ::int จำเป็น — count(*) คืน bigint ทำให้หา fn_rfm_segment(int, timestamptz) ไม่เจอ
   update public.customers c
-  set points_balance  = coalesce(s.pts, 0),
-      visit_count     = coalesce(s.visits, 0),
+  set points_balance  = coalesce(s.pts, 0)::int,
+      visit_count     = coalesce(s.visits, 0)::int,
       last_visited_at = s.last_at,
-      rfm_segment     = public.fn_rfm_segment(coalesce(s.visits, 0), s.last_at)
+      rfm_segment     = public.fn_rfm_segment(coalesce(s.visits, 0)::int, s.last_at)
   from (
     select
       customer_id,
